titanic_ml_project/requirements.txt
absl-py==0.10.0
alembic==1.4.1
astunparse==1.6.3
azure-core==1.8.2
azure-storage-blob==12.5.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
cryptography==3.1.1
cycler==0.10.0
databricks-cli==0.12.0
docker==4.3.1
entrypoints==0.3
Flask==1.1.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.9
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gorilla==0.3.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Mako==1.1.3
Markdown==3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mlflow==1.11.0
msrest==0.6.19
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.3
patsy==0.5.1
Pillow==7.2.0
prometheus-client==0.8.0
prometheus-flask-exporter==0.18.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
pywin32==227
PyYAML==5.3.1
querystring-parser==1.2.4
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
shellingham==1.3.2
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.13
sqlparse==0.3.1
statsmodels==0.12.0
tabulate==0.8.7
tenacity==6.2.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
typer==0.3.2
urllib3==1.25.10
waitress==1.4.4
websocket-client==0.57.0
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
